requirements.txt
# Dependencies cho Bot Du Lịch Núi Bà Đen
python-zalo-bot>=1.0.0
python-dotenv>=0.19.0
requests>=2.28.0
aiohttp>=3.8.0
supabase>=1.0.0
google-generativeai>=0.3.0
pytz>=2023.3
asyncio
logging